backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
web3==6.11.3
pydantic==2.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23